% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/station.R
\name{ts_add_stations}
\alias{ts_add_stations}
\title{Add Station}
\usage{
ts_add_stations(stations, file = getOption("tsdbr.file", "tscbh.sqlite"))
}
\arguments{
\item{stations}{A data frame of stations with columns Station, Parameter,
Period and Site. The optional columns are
LowerLimit, UpperLimit, Elevation, StationName and Comments.}

\item{file}{A string of the name of the database file.}
}
\value{
The imported station data.
}
\description{
Add Station
}
