% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad.R
\name{ts_add_triad}
\alias{ts_add_triad}
\title{Add Triad}
\usage{
ts_add_triad(child, parent1, parent2, conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{child}{A string of the child station.}

\item{parent1}{A string of the first parent station.}

\item{parent2}{A string of the second parent station.}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
A data frame of the imported triad.
}
\description{
Add Triad
}
