% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triad.R
\name{ts_add_triads}
\alias{ts_add_triads}
\title{Add Triads}
\usage{
ts_add_triads(triads, conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{triads}{A data frame of triads with columns Child, Parent1, Parent2.}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
The imported triad data.
}
\description{
Add Triads
}
