% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{ts_create_db}
\alias{ts_create_db}
\title{Create Time-Series Database}
\usage{
ts_create_db(file)
}
\arguments{
\item{file}{A string of the name of the database file.}
}
\value{
A connection to the database.
}
\description{
Create Time-Series Database
}
