% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ts_get_data}
\alias{ts_get_data}
\title{Get Data}
\usage{
ts_get_data(stations = ts_get_stations()$Station, start_date = end_date -
  366L, end_date = Sys.Date(), period = "hour", na_rm = FALSE,
  status = "reasonable", file = getOption("tsdbr.file", "ts.db"))
}
\arguments{
\item{stations}{A character vector of the stations.}

\item{start_date}{The start date.}

\item{end_date}{The end date.}

\item{period}{A string of the period to aggregate and average by.
The possible values are 'year', 'month', 'day', 'hour'.}

\item{na_rm}{A flag indicating whether to remove missing values (if possible) when aggregating.}

\item{status}{A string of the worse type of data to get.
The possible values are 'reasonable', 'questionable' or 'erroneous'.}

\item{file}{A string of the name of the database file.}
}
\value{
A data frame of the requested data.
}
\description{
Get Data
}
