% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-bch.R
\name{ts_read_bch}
\alias{ts_read_bch}
\title{Read BC Hydro Data File}
\usage{
ts_read_bch(file = "tscbh.xls")
}
\arguments{
\item{file}{A string specifying the path to the file.}
}
\description{
Reads in an Excel spreadsheet of BC Hydro data as provided by Power Records.
The station cannot be consistently determined from the file so must be
manually added.
}
