% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-brd.R
\name{ts_read_brd}
\alias{ts_read_brd}
\title{Read Brilliant Dam Data File}
\usage{
ts_read_brd(file = "brd.csv", utc_offset = -8L)
}
\arguments{
\item{file}{A string specifying the path to the file.}

\item{utc_offset}{A integer of the utc offset which must lie between -12 and 14.}
}
\description{
A utc_offset of -8 is equivalent to Pacific Standard Time.
}
