% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-zrxp.R
\name{ts_read_zrxp}
\alias{ts_read_zrxp}
\title{Read zrxp Data File}
\usage{
ts_read_zrxp(file = "tscbh.zrxp", utc_offset = -8L)
}
\arguments{
\item{file}{A string specifying the path to the file.}

\item{utc_offset}{A integer of the utc offset which must lie between -12 and 14.}
}
\description{
A utc_offset of -8 is equivalent to Pacific Standard Time.
}
