% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{ts_write_csv}
\alias{ts_write_csv}
\title{Write csv}
\usage{
ts_write_csv(data, file = "tsdbr.csv")
}
\arguments{
\item{data}{The data frame with a minimum of columns Station and DateTime}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing.  \code{""} indicates output to the console.}
}
\value{
An invisible copy of saved data.
}
\description{
The initial columns in the csv file are as follows
\describe{
  \item{Year}{The year of the value.}
  \item{Month}{The month of the value.}
  \item{Day}{The day of the value.}
  \item{Hour}{The hour of the value.}
  \item{Station}{The station}
}
The csv file should also include the following columns
\describe{
  \item{Recorded}{The value as provided by the station maintainer.}
  \item{Corrected}{The corrected value (see below).}
  \item{Status}{The reliability of the corrected value (see below).}
  \item{Parameter}{The parameter type, ie, Discharge, Elevation, Water Temperature etc.}
  \item{Units}{The units of the value.}
  \item{StationName}{The name of the station.}
  \item{Comments}{Comments on the value.}
}
}
\details{
There are three possible status values
\describe{
  \item{Reasonable}{There is nothing obviously wrong with the value.}
  \item{Questionable}{The value may be wrong.}
  \item{Erroneous}{The value is definitely wrong.}
}
In general questionable values are those which are in the range of possible 
values but are inconsistent with spatial or temporal neighbours 
while erroneous value are those which are outside the range of possible values, ie negative discharge.

If a value is erroneous, questionable or missing (indicated by NA) 
then we will (time permitting) provide a corrected value based on neighbouring values.
Otherwise the corrected value is simply the recorded value.
The status column refers to the corrected value and 
corrected values which differ from the recorded value are coded as Questionable.

If data gaps are not important it is recommended that you only include Reasonable values.
Otherwise it is recommended that you only include Questionable values 
that you consider to be reliable based on plotting and analysis.
}
