% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ts_add_data}
\alias{ts_add_data}
\title{Add Data}
\usage{
ts_add_data(data, resolution = "abort", conn = getOption("tsdbr.conn",
  NULL))
}
\arguments{
\item{data}{A data frame of data with columns Station, DateTime, Recorded.
Additional optional columns include Corrected, Status, Comments.}

\item{resolution}{A string of the action to take with regard to existing values.
Options are 'abort', 'ignore' or 'replace'.}

\item{conn}{An object of class SQLiteConnection.}
}
\value{
A data frame of the imported parameters.
}
\description{
Add Data
}
