% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doctor.R
\name{ts_doctor_db}
\alias{ts_doctor_db}
\title{Doctor Database}
\usage{
ts_doctor_db(conn = getOption("tsdbr.conn", NULL))
}
\arguments{
\item{conn}{An object of class SQLiteConnection.}
}
\value{
A flag indicating whether or not the database passed the checks (or was fixed)
}
\description{
Doctor Database
}
