% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{ts_write_csv}
\alias{ts_write_csv}
\title{Write csv}
\usage{
ts_write_csv(data, file = "tsdbr.csv")
}
\arguments{
\item{data}{The data frame with minimum of columnms Station and DateTime}

\item{file}{either a character string naming a file or a \link{connection}
    open for writing.  \code{""} indicates output to the console.}
}
\value{
An invisible copy of saved data.
}
\description{
Write csv
}
